[OMNIML-4788] specdec_bench: merge s3_utils into upload_to_s3, drop internal defaults
upload_to_s3.py was a single-file CLI that only imported helpers from a
neighboring s3_utils.py module. The split bought nothing — there was no
second consumer, no plan for one, and the package import made the
uploader look coupled to the specdec_bench library when in fact it has
zero references to bench code (just sentinel filenames). Merging makes
this a drop-in single-file tool: anyone can `wget upload_to_s3.py` and
run it without `pip install`ing the package or fiddling with PYTHONPATH.

Also removes the hardcoded `S3_DEFAULT_ENDPOINT = "https://pdx.s8k.io"`
constant + the NVIDIA-internal team-specdec-workgroup defaults that used
to live in s3_utils. Endpoint / key id / secret now default to empty
(env-var or --flag only). This is a *public* example repo; team-specific
infrastructure URLs shouldn't ship in it.

Net: 1 file removed (-75 LOC), upload_to_s3.py grew +62 LOC, -13 overall.
Diff for downstream callers: change `from specdec_bench.s3_utils import …`
to `from upload_to_s3 import …`. The pensieve-intern submit tool (planned
follow-up) will set `--endpoint https://pdx.s8k.io --key-id …` from
team-secrets storage rather than relying on a baked-in default.
